Use uint16 for depths -- saves memory and could lead to potential optimizations later

+++ src/rasterizer.c
@@ -79,7 +79,7 @@ void rasterizer_draw_short_span(
             // Set the pixel
             rasterizer_set_pixel(ctx->user_ptr, ix, iy, p);
 
-            ctx->depths[ix] = z;
+            ctx->depths[ix] = (uint16_t) z;
         }
 
         z += step_z;
@@ -154,7 +154,7 @@ void rasterizer_draw_long_span(
                 c |= p << shift;
                 mask |= 3 << shift;
 
-                ctx->depths[ix] = z;
+                ctx->depths[ix] = (uint16_t) z;
             }
 
             z += step_z;
